requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
scipy>=1.9.0
pillow>=9.0.0

# Diffusion Models
diffusers>=0.21.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.7.0
timm>=0.9.0  # Vision Transformers

# Robotics Simulation (Optional - install separately if needed)
# mani-skill2>=0.5.0  # Commented out due to sapien dependency conflicts
gymnasium>=0.29.0
# gymnasium-robotics>=1.2.0  # Optional, install if needed

# Data Processing
zarr>=2.13.0
h5py>=3.8.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
wandb>=0.15.0  # Experiment tracking
tensorboard>=2.13.0

# Language Models
ftfy>=6.1.0
regex>=2022.10.31
clip @ git+https://github.com/openai/CLIP.git

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.6.0  # Tensor operations
hydra-core>=1.3.0  # Configuration management
omegaconf>=2.3.0

# Development
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
